renamed the package to be more relevant
